Limit the number of entries shown, order ascendingly by time and changed seconds to float

+++ src/models.rs
@@ -6,5 +6,5 @@ pub struct Entries {
     pub id: i32,
     pub level_id: i32,
     pub username: String,
-    pub seconds: i32,
+    pub seconds: f32,
 }
